Update readme to include installation/usage instructions.
Naming

This assumes the name of this repo is `zazu-mac-omnifocus` which signifies it's
a zazu plugin and only works on mac.

Travis

This includes a travis badge, if you want to enable travis you can go your
account preferences here:

https://travis-ci.org/profile/bookstein

Plugin Directory

You can submit this to the plugin direcotry by creating a pull request to Zazu
Core. Simply make a new markdown file in the `_packages` directory in Zazu named
`omnifocus.md` or something similar. You can copy one of the other ones for a
template.

http://zazuapp.org/documentation/plugins/#submitting-your-plugin